Refactor map component
